xacml-core/src/main/java/com/artagon/xacml/policy/type/DayTimeDurationType.java
package com.artagon.xacml.policy.type;

import javax.xml.datatype.Duration;

import com.artagon.xacml.policy.AttributeDataType;
import com.artagon.xacml.policy.BaseAttributeValue;


public interface DayTimeDurationType extends AttributeDataType
{
	DayTimeDurationValue create(Object value);
	DayTimeDurationValue fromXacmlString(String v);
	
	final class DayTimeDurationValue extends BaseAttributeValue<Duration>
	{
		public DayTimeDurationValue(DayTimeDurationType type, 
				Duration value) {
			super(type, value);
		}
	}
}

